Add reglist object describing a bridging/mirroring registration

By specifying either reference, sample or mirror arguments, you can add a bridging or mirroring registration, respectively, to the list of those in use for xform_brain and mirror_brain.

add_reglist(x, reference = NULL, sample = NULL, mirror = NULL,
  temp = TRUE, ...)

Arguments

x
A single reglist object (which )

reference, sample
The reference and sample brains (in character or templatebrain form) for a bridging registration.

mirror
The reference brain (in character or templatebrain form) for a mirroring registration.

temp
Whether to store the on disk representation in a session-specific temporary folder (that will be removed when R closes). Defaults to TRUE.

...
Additional arguments passed to saveRDS e.g. to control compression when the reglist object is saved to disk.

See also

Examples

# NOT RUN {
library(nat.flybrains)
# mirroring registration for a specific template brain object
add_reglist(mirroring, mirror=JFRC2013)
# equivalent but withhout needing to construct the template
add_reglist(mirroring, mirror="JFRC2013")

# add a bridging registration between two brains
add_reglist(bridging, reference=JFRC2, sample=JFRC2013)

# }