Standard local checkout location for extra registration directories

Standard local checkout location for extra registration directories

local_reg_dir_for_url(url = NULL)

Arguments

url

Character vector containing a url. When url=NULL defaults to giving the base path.

Value

path(s) containing registrations on disk.

Details

When called without any argument returns the root directory that will be inspected for extra registrations. You can put a sub-folder yourself there manually and then call add_reg_folders, but you are much better off in general using download_reg_repo to install from a github repository such as this one of ours:. jefferislab/BridgingRegistrations

Note that this folder will always be the same place on a machine i.e. this defines a consistent, persistent location on disk to store data across sessions. You can modify the location of this folder by editing the R_USER_DATA_DIR environment variable. While this is not recommended on a personal machine, it may be necessary on a server. See rappdirs::user_data_dir for details.

When called with a url, a SHA1 hash will be calculated for the URL and appended to the basepath. This should ensure that locations derived from different URLs do not clash.

See also