Construct templatebrain object for an image registration template

templatebrain objects encapsulate key information for the reference brain in an image registration. Usually this will be a standard template brain used for many registrations. It will normally be much more convenient to use as.templatebrain methods to convert an image file or an im3d object into a templatebrain.

templatebrain(name, regName = name, type = NULL, sex = NULL,
  dims = NULL, BoundingBox = NULL, voxdims = NULL, units = NULL,
  description = NULL, doi = NULL, ...)

Arguments

name

the full name of the template.

regName

the short name. This will be the stem used to prefix registrations (e.g. JFRC2_someimage.list) for this template brain and likely also the stem of the template brain image (e.g. JFRC2.nrrd).

type

one of c('single brain', 'average'), indicating whether the template brain has been created from just one image, or is the average of multiple images.

sex

the sex of the template brain. For templates with type=='average', the possibility of sex='intersex' exists.

dims

dimensions of the image (number of voxels).

BoundingBox

physical dimensions of the image (see boundingbox).

voxdims

physical spacing between voxels.

units

units of physical measurements (e.g. microns).

description

details of the template.

doi

a DOI for the original template brain image.

...

additional named arguments that will be added as fields to the templatebrain object.

Value

A list with class templatebrain.

Details

A variety of methods are available to work on templatebrain objects. See templatebrain-meths for basic methods. The two main functions that are availavle for using template brains are xform_brain and mirror_brain.

templatebrain objects are only useful for transformation processes when the BoundingBox is specified to define the physical extent of the volume. We use the definition of the Amira 3D visualisation and analysis software. This corresponds to the node centers option in the http://teem.sourceforge.net/nrrd/format.html. The bounding box can be obtained from NRRD or AmiraMesh format files. See boundingbox for details.

See also

as.templatebrain, templatebrain-meths, xform_brain, mirror_brain.