Plot 3D surface of a template brain

Plot 3D surface of a template brain

# S3 method for templatebrain
plot3d(x, col = "grey", alpha = 0.3, ...)

Arguments

x
the template brain to plot.

col
the color of the surface.

alpha
the alpha value of the surface.

...
extra arguments to pass to plot3d.

Details

This function will work immediately for the standard templatebrain defined in the package documentation. If passed an object called e.g. FCWB it expects to find another object named FCWB.surf containing the surface information. If you follow this naming convention for user-defined refbrains it will work for them as well.